Update to Google Storage APIs v 0.14
After commenting my findings on
[this](https://github.com/googleapis/elixir-google-api/issues/972)
issue, it turns out that there's no good way to automate uploading
binary data and it needs to be hard-coded, similar to how the other
Google Storage API clients do it. This new function is called
`storage_objects_insert_iodata` and allows a binary (string) or iolist
to be used when performing a multipart upload.